pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "omniai"
version = "0.1.0"
description = "The Sovereign AI Foundation Layer for Emerging Economies"
readme = "README.md"
license = "MIT"                   # ✅ SPDX short identifier (string, NOT table)
license-files = ["LICENSE"]       # ✅ Explicit license file
authors = [
    { name = "Antony Henry Oduor Onyango", email = "harryoduwor@gmail.com" }
]
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
]
requires-python = ">=3.10"
dependencies = [
    # Web & API
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    
    # Data Validation & Settings
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",        # ← NEW: for config management
    
    # Database
    "sqlalchemy[asyncio]>=2.0.36",     # ← NEW: core ORM + async support
    "asyncpg>=0.29.0",                 # ← NEW: async PostgreSQL driver
    
    # Security & Auth
    "python-jose[cryptography]>=3.3.0", # ← NEW: JWT encoding/decoding
    "passlib[bcrypt]>=1.7.4",          # ← NEW: password hashing
    "bcrypt>=4.0.0",                    # Explicitly add bcrypt
    
    # HTTP Client (for future service calls)
    "httpx>=0.28.0",
    
    # Structured Logging (optional but recommended)
    "structlog>=24.4.0",               # ← NEW: production-grade logs
    "email-validator>=2.2.0",
    "python-multipart>=0.0.9",
    "charset-normalizer>=3.3.0",
    
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0",
    "mypy>=1.13.0",
    "ruff>=0.8.0",
    "bandit>=1.7.9",                   # ← NEW: security linter
    "httpx>=0.28.0",                   # already in main, but safe to repeat
    "python-dotenv>=1.0.1",
    "safety>=2.3.0,<3.0",
    
]

[tool.setuptools.packages.find]
where = ["src"]
include = ["omniai*"]

[tool.pytest.ini_options]
asyncio_mode = "auto"  # or "strict" — but if "strict", you MUST use @pytest.mark.asyncio
cache_dir = "/tmp/.pytest_cache"

# This ensures pytest always looks in tests/, even if run from weird directories.
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]

[tool.ruff]
lint.select = ["E", "W", "F", "I", "B", "SIM", "TID", "C4"]
lint.ignore = ["E501"]  # allow long lines
line-length = 88

[tool.ruff.lint.per-file-ignores]
"src/omniai/tests/unit/test_auth.py" = ["B008"]
"src/omniai/api/v1/me.py" = ["B008"]
"src/omniai/api/v1/auth.py" = ["B008"]